should of been in the last commit.
git-svn-id: svn://fly.thruhere.net/forsaken@1557 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/screenpolys.c
@@ -1174,7 +1174,7 @@ void UpdateCountdownDigits( void )
 	}
 
 	Countdown_Float -= Time_Diff;
-	Number = (LONGLONG) Countdown_Float;
+	Number = (int) Countdown_Float;
 
 	if( Countdown_Float < 0.0F )
 	{
